Tons of Fixes
I fixed about 25 warnings. I created a whole new kind of Segue, which
dismisses the current view controller. Before, we were simply
presenting the old view controller on top of the current one, which was
incorrect and would ultimately lead to crashes as well as ugly
animations. I also went ahead and fixed every “back” button to work
with this new method of dismissing. I also update the username and
password textfields to display white text to match the rest of the UI.
I also updated image views throughout the app to fix the “stretching”
issue that was present in previous builds.